+++ src/pages/StudyDetailPage/components/Emoji/EmojiPopOver/EmojiAdd/EmojiAdd.jsx
@@ -26,7 +26,11 @@ const EmojiAdd = ({ emojis, setEmojis, id, updateEmoji, toastHandler }) => {
 
       setEmojis((prev) => [...prev, newEmoji]);
     } else if (!selectEmoji && emojiRef.current === e.emoji) {
-      toastHandler('emoji', '이모지 등록중입니다! 잠시 후 다시 시도해주세요!');
+      toastHandler(
+        'error',
+        '이모지 등록중입니다! 잠시 후 다시 시도해주세요!',
+        'emoji',
+      );
     } else {
       //update emoji id 같이
       updateEmoji(selectEmoji.id);
